refactor files structure
also introduces RootRenderObjectMixin to move RenderView to bindings file since scheduleInitialLayout uses private fields
