+++ framework.dart
@@ -1,4 +1,7 @@
 library framework;
 
 export 'bindings.dart';
-export 'trees.dart';
+export 'geometry.dart';
+export 'render_object.dart';
+export 'widget.dart';
+export 'element.dart';
